moved rewrite algorithm into separate file
